refactor: collapse to one eager registry per domain, self-split heavy deps
Replace the dual eager/lazy registries (and the key-set pinning that kept
the twins in sync) with a single flat registry: one plugin per domain,
every component eager. The two components that pull a heavy external
dependency now split it from inside the component, following the existing
LazyDatePickerControl pattern:

- chart -> chart-view (Recharts)
- rich-editor -> rich-editor-field (TipTap)

The registry never has to know a component is expensive: small ones are
eager, and a heavy one code-splits its own dependency. Table is eager (no
heavy external dep). Removes the three eager.ts twins, registry/{eager,
lazy}.ts, the eagerRegistry export + its package.json subpaths, and the
now-unused form skeleton. Renames chatPlugin/notificationsPlugin to
*Components, aligns chat's plugin name and the layout import path. Build
confirms Recharts and TipTap still emit as separate chunks.
