+++ resources/js/chat/index.ts
@@ -3,8 +3,8 @@ import ChatBoxComponent from "./components/chat-box";
 import { TextPart } from "./parts/text";
 import { ToolCallPart } from "./parts/tool-call";
 
-export const chatPlugin = createPlugin({
-  name: "chat",
+export const chatComponents = createPlugin({
+  name: "lattice/chat",
   components: {
     "chat.box": eagerComponent(ChatBoxComponent),
     "chat.part.text": eagerComponent(TextPart),
