--- resources/js/form/eager.ts
@@ -1,49 +1,0 @@
-import {
-  type ComponentRegistration,
-  createPlugin,
-  eagerComponent,
-} from "@lattice-php/lattice/core/registry";
-import type { FormComponentType } from "./index";
-import {
-  BuilderComponent,
-  CheckboxComponent,
-  ChoiceComponent,
-  DateInputComponent,
-  DateTimeInputComponent,
-  FileUploadComponent,
-  FormComponent,
-  HiddenInputComponent,
-  NumberInputComponent,
-  OtpInputComponent,
-  PasswordInputComponent,
-  RepeaterComponent,
-  SelectComponent,
-  TextareaComponent,
-  TextInputComponent,
-  TimeInputComponent,
-  ToggleComponent,
-} from "./components";
-import { RichEditorComponent } from "./components/fields/rich-editor";
-
-const components = {
-  form: eagerComponent(FormComponent),
-  "field.builder": eagerComponent(BuilderComponent),
-  "field.checkbox": eagerComponent(CheckboxComponent),
-  "field.choice": eagerComponent(ChoiceComponent),
-  "field.date-input": eagerComponent(DateInputComponent),
-  "field.date-time-input": eagerComponent(DateTimeInputComponent),
-  "field.file-upload": eagerComponent(FileUploadComponent),
-  "field.hidden-input": eagerComponent(HiddenInputComponent),
-  "field.number-input": eagerComponent(NumberInputComponent),
-  "field.otp": eagerComponent(OtpInputComponent),
-  "field.password-input": eagerComponent(PasswordInputComponent),
-  "field.repeater": eagerComponent(RepeaterComponent),
-  "field.rich-editor": eagerComponent(RichEditorComponent),
-  "field.select": eagerComponent(SelectComponent),
-  "field.textarea": eagerComponent(TextareaComponent),
-  "field.text-input": eagerComponent(TextInputComponent),
-  "field.time-input": eagerComponent(TimeInputComponent),
-  "field.toggle": eagerComponent(ToggleComponent),
-} satisfies Record<FormComponentType, ComponentRegistration>;
-
-export const eagerFormComponents = createPlugin({ components, name: "lattice/form" });
