+++ resources/js/form/index.ts
@@ -1,74 +1,45 @@
+import { createPlugin, eagerComponent } from "@lattice-php/lattice/core/registry";
 import {
-  type ComponentRegistry,
-  createPlugin,
-  lazyComponent,
-} from "@lattice-php/lattice/core/registry";
-import type { RendererComponent, RendererComponentModule } from "@lattice-php/lattice/core/types";
-import { FormSkeletonComponent } from "./skeleton";
+  BuilderComponent,
+  CheckboxComponent,
+  ChoiceComponent,
+  DateInputComponent,
+  DateTimeInputComponent,
+  FileUploadComponent,
+  FormComponent,
+  HiddenInputComponent,
+  NumberInputComponent,
+  OtpInputComponent,
+  PasswordInputComponent,
+  RepeaterComponent,
+  SelectComponent,
+  TextareaComponent,
+  TextInputComponent,
+  TimeInputComponent,
+  ToggleComponent,
+} from "./components";
+import { RichEditorComponent } from "./components/fields/rich-editor";
 
-type FormComponentName =
-  | "BuilderComponent"
-  | "CheckboxComponent"
-  | "ChoiceComponent"
-  | "DateInputComponent"
-  | "DateTimeInputComponent"
-  | "FileUploadComponent"
-  | "FormComponent"
-  | "HiddenInputComponent"
-  | "NumberInputComponent"
-  | "OtpInputComponent"
-  | "PasswordInputComponent"
-  | "RepeaterComponent"
-  | "SelectComponent"
-  | "TextareaComponent"
-  | "TextInputComponent"
-  | "TimeInputComponent"
-  | "ToggleComponent";
-
-const loadFormComponents = () => import("./components");
-
-function loadFormComponent<TType extends string>(
-  name: FormComponentName,
-): () => Promise<RendererComponentModule<TType>> {
-  return async () => {
-    const components = await loadFormComponents();
-
-    return {
-      default: components[name] as unknown as RendererComponent<TType>,
-    };
-  };
-}
-
-const components = {
-  form: lazyComponent(loadFormComponent("FormComponent"), {
-    fallback: FormSkeletonComponent,
-  }),
-  "field.builder": lazyComponent(loadFormComponent("BuilderComponent")),
-  "field.checkbox": lazyComponent(loadFormComponent("CheckboxComponent")),
-  "field.choice": lazyComponent(loadFormComponent("ChoiceComponent")),
-  "field.date-input": lazyComponent(loadFormComponent("DateInputComponent")),
-  "field.date-time-input": lazyComponent(loadFormComponent("DateTimeInputComponent")),
-  "field.file-upload": lazyComponent(loadFormComponent("FileUploadComponent")),
-  "field.hidden-input": lazyComponent(loadFormComponent("HiddenInputComponent")),
-  "field.number-input": lazyComponent(loadFormComponent("NumberInputComponent")),
-  "field.otp": lazyComponent(loadFormComponent("OtpInputComponent")),
-  "field.password-input": lazyComponent(loadFormComponent("PasswordInputComponent")),
-  "field.repeater": lazyComponent(loadFormComponent("RepeaterComponent")),
-  // Loaded from its own module so TipTap is split into a separate chunk,
-  // only fetched on pages that actually render a rich editor.
-  "field.rich-editor": lazyComponent<"field.rich-editor">(async () => {
-    const { RichEditorComponent } = await import("./components/fields/rich-editor");
-
-    return { default: RichEditorComponent };
-  }),
-  "field.select": lazyComponent(loadFormComponent("SelectComponent")),
-  "field.textarea": lazyComponent(loadFormComponent("TextareaComponent")),
-  "field.text-input": lazyComponent(loadFormComponent("TextInputComponent")),
-  "field.time-input": lazyComponent(loadFormComponent("TimeInputComponent")),
-  "field.toggle": lazyComponent(loadFormComponent("ToggleComponent")),
-} satisfies ComponentRegistry;
-
-export const formComponents = createPlugin({ components, name: "lattice/form" });
-
-/** The component types this domain registers; its eager twin is pinned to this. */
-export type FormComponentType = keyof typeof components;
+export const formComponents = createPlugin({
+  components: {
+    form: eagerComponent(FormComponent),
+    "field.builder": eagerComponent(BuilderComponent),
+    "field.checkbox": eagerComponent(CheckboxComponent),
+    "field.choice": eagerComponent(ChoiceComponent),
+    "field.date-input": eagerComponent(DateInputComponent),
+    "field.date-time-input": eagerComponent(DateTimeInputComponent),
+    "field.file-upload": eagerComponent(FileUploadComponent),
+    "field.hidden-input": eagerComponent(HiddenInputComponent),
+    "field.number-input": eagerComponent(NumberInputComponent),
+    "field.otp": eagerComponent(OtpInputComponent),
+    "field.password-input": eagerComponent(PasswordInputComponent),
+    "field.repeater": eagerComponent(RepeaterComponent),
+    "field.rich-editor": eagerComponent(RichEditorComponent),
+    "field.select": eagerComponent(SelectComponent),
+    "field.textarea": eagerComponent(TextareaComponent),
+    "field.text-input": eagerComponent(TextInputComponent),
+    "field.time-input": eagerComponent(TimeInputComponent),
+    "field.toggle": eagerComponent(ToggleComponent),
+  },
+  name: "lattice/form",
+});
