+++ resources/js/notifications/index.ts
@@ -1,7 +1,7 @@
 import { createPlugin, eagerComponent } from "@lattice-php/lattice/core/registry";
 import NotificationsComponent from "./components/notifications";
 
-export const notificationsPlugin = createPlugin({
+export const notificationsComponents = createPlugin({
   name: "lattice/notifications",
   components: {
     notifications: eagerComponent(NotificationsComponent),
